refactor: create quiz creation component
